executes the cc pre-processor on input files
the pre-processor is now (poorly) invoked on the input files. the compiler will call out to cc and run the pre-processor which will produce a pre-processed copy of the input file with a .ppc suffix. For example, main.c will produce a main.c.ppc file. These files are removed after the compiler reads the file into memory.

As of writing this, the compiler can _nearly_ lex itself except for some weird bug concerning single quotes. None of the lexer is compliant to the c99 spec thus far.